Added userstream.py, needs work, see below:
So mainly right now the features i'm thinking of for birdyBot have to do with DMing users on a linux system.
To do that obviously, we'll need the streaming thing to work, which, it does but i'm not sure what dict or list tweepy is putting direct message data out to.
I don't want to do my own parsing if I don't have to, and since I know tweepy is, I'd just like to use that.